+++ protected/humhub/modules/content/widgets/richtext/converter/RichTextToHtmlConverter.php
@@ -38,12 +38,17 @@ class RichTextToHtmlConverter extends BaseRichTextConverter
     /**
      * @var string HtmlPurifier HTML.Allowed configuration
      */
-    public $htmlAllowed = 'h1,h2,h3,h4,h5,h6,br,b,i,strong,em,a,pre,code,img,tt,div,ins,del,sup,sub,p,ol,ul,table,thead,tbody,tfoot,blockquote,dl,dt,dd,kbd,q,samp,var,hr,li,tr,td,th,s,strike';
+    public $htmlAllowed = 'h1,h2,h3,h4,h5,h6,br,b,i,strong,em,a,pre,code,img,video,audio,tt,div,ins,del,sup,sub,p,ol,ul,table,thead,tbody,tfoot,blockquote,dl,dt,dd,kbd,q,samp,var,hr,li,tr,td,th,s,strike';
 
     /**
      * @var string HtmlPurifier HTML.AllowedAttributes configuration
      */
-    public $htmlAllowedAttributes = 'img.src,img.alt,img.title,img.width,img.height,img.class,img.style,code.class,a.rel,a.target,a.href,a.title,th.align,td.align,ol.start';
+    public $htmlAllowedAttributes = 'img.src,img.alt,img.title,img.width,img.height,img.class,img.style,'
+        . 'video.src,video.title,video.width,video.height,video.class,video.style,video.controls,video.autoplay,video.muted,video.loop,'
+        . 'audio.src,audio.title,audio.class,audio.style,audio.controls,audio.autoplay,audio.muted,audio.loop,'
+        . 'code.class,'
+        . 'a.rel,a.target,a.href,a.title,'
+        . 'th.align,td.align,ol.start';
 
     /**
      * @var bool whether the output should be purified
@@ -88,6 +93,24 @@ class RichTextToHtmlConverter extends BaseRichTextConverter
 
             $htmlDefinition->addAttribute('a', 'target', 'Text');
             $htmlDefinition->addAttribute('img', 'style', 'Text');
+            $htmlDefinition->addElement('video', 'Block', 'Optional: (source, Flow) | Flow', 'Common', [
+                'src'      => 'URI',
+                'title'    => 'Text',
+                'width'    => 'Length',
+                'height'   => 'Length',
+                'controls' => 'Bool',
+                'autoplay' => 'Bool',
+                'muted'    => 'Bool',
+                'loop'     => 'Bool',
+            ]);
+            $htmlDefinition->addElement('audio', 'Block', 'Optional: (source, Flow) | Flow', 'Common', [
+                'src'      => 'URI',
+                'title'    => 'Text',
+                'controls' => 'Bool',
+                'autoplay' => 'Bool',
+                'muted'    => 'Bool',
+                'loop'     => 'Bool',
+            ]);
         });
     }